Generics and other fun stuff
changed the format of the template span ids to be "#NAME#" to aid in
parsing html looking for them. Now working towards getting the UI to
poll a given page and list any found template spans and present the user
with them.
making some of the functions generic to aid this as well.